Python 3.10 compatibility.
The asyncio module has deprecated the concept of a current - but not running - event loop. So `get_event_loop` has been replaced with `new_event_loop`.
This also make the `stream` interfaces unworkable, because the coroutines must share an event loop. They have been replaced with a similar `tasks` object.